ORIGIN work http://domain1 http://domain2 now
Previous patch would only allow one domain to connect to the project. This patch allow to use any domain you would like to connect to your app. example $io->origins('http://domain1 http://domain2'); or $io->origins('http://domain1:* http://domain2:*'); ..

This patch also add a suitable 403 http code when a domain tries to connect and is not in the allowed domains..

Which is better also for debugging, instead of 400 bad request...